backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ririi-q <ririi.q.27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"
pytest = "^8.3.3"
pytest-flask = "^1.3.0"
fastapi = "^0.115.2"
fastapi-oauth2 = "^1.3.0"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
pydantic = {extras = ["email"], version = "^2.9.2"}
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-multipart = "^0.0.12"
httpx = "^0.27.2"
authlib = "^1.3.2"
social-auth-core = "^4.5.4"

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []
line-length = 120

[tool.ruff.per-file-ignores]
"**/api/*.py" = ["B008"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
